@@ -1,113 +1,154 @@
-﻿using System.Collections.Concurrent;
+﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
-using Newtonsoft.Json;
+using System.Text.Json;
+using System.Threading.Tasks;
+using MessagePack;
 
 
-namespace HyperVectorDB {
-    public class HyperVectorDBIndex {
+namespace HyperVectorDB
+{
+    public class HyperVectorDBIndex
+    {
         public readonly string Name;
+
+        public int Count
+        {
+            get { return documents.Count; }
+        }
         private List<double[]> vectors;
         private List<HVDBDocument> documents;
         private readonly Dictionary<double[], HVDBQueryResult> queryCacheCosineSimilarity;
 
-        public HyperVectorDBIndex(string name) {
+        private bool fileValid = false;
+
+        private MessagePackSerializerOptions options = MessagePackSerializerOptions.Standard
+            .WithSecurity(MessagePackSecurity.UntrustedData)
+            .WithCompression(MessagePackCompression.Lz4BlockArray);
+
+        public HyperVectorDBIndex(string name)
+        {
             this.vectors = new List<double[]>();
             this.documents = new List<HVDBDocument>();
             this.queryCacheCosineSimilarity = new Dictionary<double[], HVDBQueryResult>();
             Name = name;
         }
-        public void Save(string path) {
-            //TODO: This could be done better
+        public async void Save(string path)
+        {
+            if (fileValid) { return; }
             var savepath = Path.Combine(path, Name);
-            if (!Directory.Exists(savepath)) {
+            if (!Directory.Exists(savepath))
+            {
                 Directory.CreateDirectory(savepath);
             }
-            var js = new JsonSerializer();
-            using (var sw = new StreamWriter(Path.Combine(savepath, "vectors.json"),false)) {
-                using var jw = new JsonTextWriter(sw);
-                js.Serialize(jw, vectors);
-            }
-            using (var sw = new StreamWriter(Path.Combine(savepath, "documents.json"),false)) {
-                using var jw = new JsonTextWriter(sw);
-                js.Serialize(jw, documents);
-            }
 
+            byte[] vectorsBytes = MessagePackSerializer.Serialize(vectors);
+            var vectorsToken = System.IO.File.WriteAllBytesAsync(Path.Combine(savepath, "vectors.bin"), vectorsBytes);
+
+            byte[] documentsBytes = MessagePackSerializer.Serialize(documents);
+            var documentsToken = System.IO.File.WriteAllBytesAsync(Path.Combine(savepath, "documents.bin"), documentsBytes);
+            await vectorsToken;
+            await documentsToken;
+
+            fileValid = true;
         }
-        public void Load(string path) {
-            //TODO: This could be done better
+        public async void Load(string path)
+        {
             var loadpath = Path.Combine(path, Name);
-            if (!Directory.Exists(loadpath)) {
+            if (!Directory.Exists(loadpath))
+            {
                 throw new DirectoryNotFoundException($"Directory {loadpath} not found.");
             }
-            var js = new JsonSerializer();
-            using (var sr = new StreamReader(Path.Combine(loadpath, "vectors.json"))) {
-                using var jr = new JsonTextReader(sr);
-                vectors = js.Deserialize<List<double[]>>(jr);
-            }
-            using (var sr = new StreamReader(Path.Combine(loadpath, "documents.json"))) {
-                using var jr = new JsonTextReader(sr);
-                documents = js.Deserialize<List<HVDBDocument>>(jr);
-            }
+
+            var vectorsToken = System.IO.File.ReadAllBytesAsync(Path.Combine(loadpath, "vectors.bin"));
+            var documentsToken = System.IO.File.ReadAllBytesAsync(Path.Combine(loadpath, "documents.bin"));
+
+            await vectorsToken;
+            vectors = MessagePackSerializer.Deserialize<List<double[]>>(vectorsToken.Result, options);
+
+            await documentsToken;
+            documents = MessagePackSerializer.Deserialize<List<HVDBDocument>>(documentsToken.Result, options);
+
+            fileValid = true;
         }
-        public void Add(double[] vector, HVDBDocument doc) {
-            if (vector == null) {
+        public void Add(double[] vector, HVDBDocument doc)
+        {
+            if (vector == null)
+            {
                 throw new ArgumentNullException(nameof(vector));
             }
-            if (doc == null) {
+            if (doc == null)
+            {
                 throw new ArgumentNullException(nameof(doc));
             }
-            if (vector.Length == 0) {
+            if (vector.Length == 0)
+            {
                 throw new ArgumentException("Vector length cannot be zero.", nameof(vector));
             }
             vectors.Add(vector);
             documents.Add(doc);
             ResetCaches();
         }
-        public void Remove(HVDBDocument doc) {
-            if (doc == null) {
+        public void Remove(HVDBDocument doc)
+        {
+            if (doc == null)
+            {
                 throw new ArgumentNullException(nameof(doc));
             }
             int index = documents.IndexOf(doc);
-            if (index == -1) {
+            if (index == -1)
+            {
                 throw new ArgumentException("Document not found.", nameof(doc));
             }
             vectors.RemoveAt(index);
             documents.RemoveAt(index);
             ResetCaches();
         }
-        public void Remove(double[] vector) {
-            if (vector == null) {
+        public void Remove(double[] vector)
+        {
+            if (vector == null)
+            {
                 throw new ArgumentNullException(nameof(vector));
             }
             int index = vectors.IndexOf(vector);
-            if (index == -1) {
+            if (index == -1)
+            {
                 throw new ArgumentException("Vector not found.", nameof(vector));
             }
             vectors.RemoveAt(index);
             documents.RemoveAt(index);
             ResetCaches();
         }
-        public void RemoveAt(int index) {
-            if (index < 0 || index >= documents.Count) {
+        public void RemoveAt(int index)
+        {
+            if (index < 0 || index >= documents.Count)
+            {
                 throw new ArgumentOutOfRangeException(nameof(index));
             }
             vectors.RemoveAt(index);
             documents.RemoveAt(index);
             ResetCaches();
         }
-        public void Clear() {
+        public void Clear()
+        {
             vectors.Clear();
             documents.Clear();
             ResetCaches();
         }
-        public void ResetCaches() {
+        public void ResetCaches()
+        {
             queryCacheCosineSimilarity.Clear();
+            fileValid = false;
         }
-        private HVDBQueryResult? TryHitCacheCosineSimilarity(double[] queryVector, int topK = 5) {
-            if (queryCacheCosineSimilarity.TryGetValue(queryVector, out HVDBQueryResult? result) && result.Documents.Count >= topK) {
+        private HVDBQueryResult? TryHitCacheCosineSimilarity(double[] queryVector, int topK = 5)
+        {
+            if (queryCacheCosineSimilarity.TryGetValue(queryVector, out HVDBQueryResult? result) && result.Documents.Count >= topK)
+            {
                 if (result is null) return null; //Sanity check
-                if (result.Documents.Count > topK) {
+                if (result.Documents.Count > topK)
+                {
                     result.Documents = result.Documents.Take(topK).ToList();
                     result.Distances = result.Distances.Take(topK).ToList();
                 }
@@ -115,12 +156,14 @@ namespace HyperVectorDB {
             }
             return null;
         }
-        public HVDBQueryResult QueryCosineSimilarity(double[] queryVector, int topK = 5) {
+        public HVDBQueryResult QueryCosineSimilarity(double[] queryVector, int topK = 5)
+        {
             if (queryVector == null) throw new ArgumentNullException(nameof(queryVector));
             if (topK <= 0) throw new ArgumentException("Number of results requested (k) must be greater than zero.", nameof(topK));
             // first check _tryCache
             HVDBQueryResult? cachedResult = TryHitCacheCosineSimilarity(queryVector, topK);
-            if (cachedResult != null) {
+            if (cachedResult != null)
+            {
                 return cachedResult;
             }
             var similarities = new ConcurrentBag<KeyValuePair<HVDBDocument, double>>();
@@ -128,7 +171,7 @@ namespace HyperVectorDB {
             {
                 //double similarity = 1 - Distance.Cosine(queryVector, vectors[i]);
 
-                double similarity = 1 - GalaxyBrainedMathsLOL.CosineSimilarity(queryVector, vectors[i]);
+                double similarity = 1 - Math.CosineSimilarity(queryVector, vectors[i]);
                 similarities.Add(new KeyValuePair<HVDBDocument, double>(documents[i], similarity));
             });
             var orderedData = similarities
@@ -142,7 +185,8 @@ namespace HyperVectorDB {
             );
         }
 
-        public HVDBQueryResult QueryJaccardDissimilarity(double[] queryVector, int topK = 5) {
+        public HVDBQueryResult QueryJaccardDissimilarity(double[] queryVector, int topK = 5)
+        {
             //TODO: THIS NEEDS TESTING
             //TODO: This needs caching
             if (queryVector == null) throw new ArgumentNullException(nameof(queryVector));
@@ -150,7 +194,7 @@ namespace HyperVectorDB {
             var similarities = new ConcurrentBag<KeyValuePair<HVDBDocument, double>>();
             Parallel.For(0, vectors.Count, i =>
             {
-                double similarity = 1 - GalaxyBrainedMathsLOL.JaccardDissimilarity(queryVector, vectors[i]);
+                double similarity = 1 - Math.JaccardDissimilarity(queryVector, vectors[i]);
                 similarities.Add(new KeyValuePair<HVDBDocument, double>(documents[i], similarity));
             });
             var orderedData = similarities
@@ -163,7 +207,8 @@ namespace HyperVectorDB {
               );
         }
 
-        public HVDBQueryResult QueryEuclideanDistance(double[] queryVector, int topK = 5) {
+        public HVDBQueryResult QueryEuclideanDistance(double[] queryVector, int topK = 5)
+        {
             //TODO: THIS NEEDS TESTING
             //TODO: This needs caching
             if (queryVector == null) throw new ArgumentNullException(nameof(queryVector));
@@ -171,7 +216,7 @@ namespace HyperVectorDB {
             var similarities = new ConcurrentBag<KeyValuePair<HVDBDocument, double>>();
             Parallel.For(0, vectors.Count, i =>
             {
-                double similarity = 1 - GalaxyBrainedMathsLOL.EuclideanDistance(queryVector, vectors[i]);
+                double similarity = 1 - Math.EuclideanDistance(queryVector, vectors[i]);
                 similarities.Add(new KeyValuePair<HVDBDocument, double>(documents[i], similarity));
             });
             var orderedData = similarities
@@ -184,7 +229,8 @@ namespace HyperVectorDB {
                 );
         }
 
-        public HVDBQueryResult QueryManhattanDistance(double[] queryVector, int topK = 5) {
+        public HVDBQueryResult QueryManhattanDistance(double[] queryVector, int topK = 5)
+        {
             //TODO: THIS NEEDS TESTING
             //TODO: This needs caching
             if (queryVector == null) throw new ArgumentNullException(nameof(queryVector));
@@ -192,7 +238,7 @@ namespace HyperVectorDB {
             var similarities = new ConcurrentBag<KeyValuePair<HVDBDocument, double>>();
             Parallel.For(0, vectors.Count, i =>
             {
-                double similarity = 1 - GalaxyBrainedMathsLOL.ManhattanDistance(queryVector, vectors[i]);
+                double similarity = 1 - Math.ManhattanDistance(queryVector, vectors[i]);
                 similarities.Add(new KeyValuePair<HVDBDocument, double>(documents[i], similarity));
             });
             var orderedData = similarities
@@ -206,7 +252,8 @@ namespace HyperVectorDB {
 
         }
 
-        public HVDBQueryResult QueryChebyshevDistance(double[] queryVector, int topK = 5) {
+        public HVDBQueryResult QueryChebyshevDistance(double[] queryVector, int topK = 5)
+        {
             //TODO: THIS NEEDS TESTING
             //TODO: This needs caching
             if (queryVector == null) throw new ArgumentNullException(nameof(queryVector));
@@ -214,7 +261,7 @@ namespace HyperVectorDB {
             var similarities = new ConcurrentBag<KeyValuePair<HVDBDocument, double>>();
             Parallel.For(0, vectors.Count, i =>
             {
-                double similarity = 1 - GalaxyBrainedMathsLOL.ChebyshevDistance(queryVector, vectors[i]);
+                double similarity = 1 - Math.ChebyshevDistance(queryVector, vectors[i]);
                 similarities.Add(new KeyValuePair<HVDBDocument, double>(documents[i], similarity));
             });
             var orderedData = similarities
@@ -227,7 +274,8 @@ namespace HyperVectorDB {
               );
         }
 
-        public HVDBQueryResult QueryCanberraDistance(double[] queryVector, int topK = 5) {
+        public HVDBQueryResult QueryCanberraDistance(double[] queryVector, int topK = 5)
+        {
             //TODO: THIS NEEDS TESTING
             //TODO: This needs caching
             if (queryVector == null) throw new ArgumentNullException(nameof(queryVector));
@@ -235,7 +283,7 @@ namespace HyperVectorDB {
             var similarities = new ConcurrentBag<KeyValuePair<HVDBDocument, double>>();
             Parallel.For(0, vectors.Count, i =>
             {
-                double similarity = 1 - GalaxyBrainedMathsLOL.CanberraDistance(queryVector, vectors[i]);
+                double similarity = 1 - Math.CanberraDistance(queryVector, vectors[i]);
                 similarities.Add(new KeyValuePair<HVDBDocument, double>(documents[i], similarity));
             });
             var orderedData = similarities
